+++ internal/test/common/src/main/kotlin/com/github/jcornaz/miop/internal/test/TestUtils.kt
@@ -9,3 +9,5 @@ inline fun <reified T : Throwable> assertThrows(block: () -> Unit): T =
         } catch (t: Throwable) {
             t as? T ?: fail("${T::class} was expected but ${t::class} was thrown")
         } ?: fail("${T::class} was expected but no exception was thrown")
+
+expect fun runTest(block: suspend () -> Unit)