+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Add `launchConsumeEach` to launch a consumer coroutine
 * Add `SubscribableVariable.bind()` function to easily bind subscribable variables
 * Add `distinctUntilChanged` operator
+* Add `IoPool` dispatcher
 
 ### JavaFx integration
 Add `miop-javafx` module, which provide: